backend/requirements.txt
# FastAPI Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3
pydantic-settings==2.1.0

# LLM Integration
openai==1.12.0
httpx==0.24.1

# ComfyUI API Client
websocket-client==1.7.0
requests==2.31.0

# Image Processing
pillow==10.2.0

# Utilities
python-dotenv==1.0.1
aiofiles==23.2.1
